Add an individual play mode
This mode behaves as if the current player is the only person that has
ever played each levelset, ignoring whether other players have solved a
level or not.  By making trials independent we should be able to collect
more data on which factors help players solve levels.  This option
probably doesn't make sense for every game mode, but should work
properly with the base game.

Enable it with the --individual flag in publish_hits.