Add an IRB consent page
The consent page is always displayed before an experiment.  An event is
logged when the player agrees to the terms.

The logEvent client code is changed such that the jQuery deferred object
from the RPC call is returned, allowing consumers to attach thier own
success/failure handlers without necessarily having to pass them in.  We
may want to follow that pattern for all RPC calls to simplify the code.