Add semi-automated experiment control
Experiment information is displayed on the dashboard.  An experiment
configuration file must be passed to the server with the --experiments
command-line option containing requirements and command-line arguments
for each experiment.  Data about these experiments is loaded from the
logs directory, the database, and MTurk, and is used to determine
whether more hits need to be published to meet the requirements.