Fix min finished per level calculation
The min finished per level stat on the dashboard now accounts for levels
with zero finishes.  A 'lvls' parameter is supported in the experiments
file to allow experiments that are only interested in that stat for a
limited number of levels (i.e. new levels are added to the levelset that
don't count for that experiment), with the default setting counting all
levels in the levelset.