Preserve levelset ordering for level play order
The order that levels are presented to players now matches the order
that they are specified in the levelset when no other sorting criteria
are met (e.g. the number of players that have seen them.)  This also
disables counting the number of players that have seen a level in
individual mode.