Implement column swapping
If enabled with the --colSwap option, columns are swapped using the
manual swap definitions in server.py.  The least used swap for a level
is served first, with levels that were served without the option
counting as the default (0) swap.  In individual mode swaps that the
player has not seen are preferred over the global least-used metric.