Add lvlset/additional invariant loading to verify.py
If a levelset is specified with the --lvlset option then each level in
the levelset will be verified, automatically filtering invariants using
the level name.

A file containing additional invariants in CSV format (level name and a
semicolon-separated list of invariants) can be specified using the
--additionalInvs option.  Additional invariants will be included
automatically when verifying all levels in a levelset.