Implement stored verification results + automation
The --write flag enables writing results to the database for --lvlset
runs only.  If --read is specified then results will be read from the
database before verification begins to avoid processing levels that have
already been proven with the given configuration.  The --auto flag
enables processing in a loop levels can be automatically re-verified as
new invariants come in.