Add an hourly HIT limit for automation
This prevents the number of HITs created within an hour from exceeding
the limit specified in the experiment configuration.  The HIT creation
time is pulled from MTurk, so data will still be correct if the server
needs to be restarted.  There is still a race condition for HIT creation
if multiple servers are running.